+++ CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG
 * fixed plotting ideograms of attributes other than age
 * improved database startup tests
 * fixed plateau age calculation
+* fixed spectrum label overlay. layout required when component's padding changes